refactor(jsii): refactored to decrease the timing of execution
